chatbot/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0
requests>=2.28.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.21.0
PyPDF2>=3.0.0
python-docx>=0.8.0
